Added Apache POI to parse excel file. However it turns out the Excel file is not an Excel file. It is actually a html file. so I'll have to figure out how to parse it tomorrow.
Can't use txt file because in the excel file it will show N/A for values
it doesn't have. But in the txt file it shows nothing, which leads to
shorter List<Strings> than expected, and it would be hard to use fixed
width to import the info. So parsing the HTML version of the list should
be better.